fix(content): remove lorem ipsum
