+++ libs/pages/environments/src/lib/ui/help-sidebar/help-sidebar.spec.tsx
@@ -1,0 +1,10 @@
+import { render } from '@testing-library/react'
+
+import HelpSidebar from './help-sidebar'
+
+describe('HelpSidebar', () => {
+  it('should render successfully', () => {
+    const { baseElement } = render(<HelpSidebar />)
+    expect(baseElement).toBeTruthy()
+  })
+})
