+++ libs/pages/environments/src/lib/ui/page-edit-deployment-rule/page-edit-deployment-rule.tsx
@@ -19,6 +19,7 @@ import { Control, Controller } from 'react-hook-form'
 import { useState } from 'react'
 import { Cluster } from 'qovery-typescript-axios'
 import { Value } from '@console/shared/interfaces'
+import HelpSidebar from '../help-sidebar/help-sidebar'
 
 export interface PageEditDeploymentRuleProps {
   listHelpfulLinks: BaseLink[]
@@ -87,13 +88,6 @@ export function PageEditDeploymentRule(props: PageEditDeploymentRuleProps) {
       label: 'SUNDAY',
       value: 'SUNDAY',
     },
-  ]
-
-  const LIST = [
-    'An environment help you to amet minim mollit non deserunt.',
-    'Ullamco est sit aliqua dolor do amet sint.',
-    'Velit officia consequat duis enim velit mollit.',
-    'Exercitation veniam consequat sunt nostrud amet.',
   ]
 
   const navigate = useNavigate()
@@ -363,43 +357,7 @@ export function PageEditDeploymentRule(props: PageEditDeploymentRuleProps) {
         </div>
         <HelpSection description="Need help? You may find these links useful" links={listHelpfulLinks}></HelpSection>
       </div>
-      <div className="w-right-help-sidebar border-l border-element-light-lighter-400">
-        <div className="p-10 border-b border-element-light-lighter-400">
-          <span className="text-4xl" role="img" aria-label="light">
-            💡
-          </span>
-          <h2 className="h5 text-text-700 mt-5 mb-5">What is an organization, what is a project?</h2>
-          <ul className="text-sm ml-2">
-            {LIST.map((l, index) => (
-              <li
-                className="text-text-500 mb-2 flex gap-3 before:content-[''] before:w-1 before:h-1 before:rounded-full before:shrink-0 before:mt-2 before:bg-text-500"
-                key={index}
-              >
-                {l}
-              </li>
-            ))}
-          </ul>
-        </div>
-        <div className="p-10">
-          <p className="text-sm text-text-500 mb-5">You may find these links useful</p>
-          <a
-            href="https://hub.qovery.com/docs/using-qovery/configuration/environment/"
-            target="_blank"
-            rel="noreferrer"
-            className="link text-accent2-500 text-sm block mb-3"
-          >
-            How to configure an environment <Icon name="icon-solid-arrow-up-right-from-square" />
-          </a>
-          <a
-            href="https://hub.qovery.com/docs/using-qovery/configuration/environment/"
-            target="_blank"
-            rel="noreferrer"
-            className="link text-accent2-500 text-sm block"
-          >
-            Set parameters on my environment <Icon name="icon-solid-arrow-up-right-from-square" />
-          </a>
-        </div>
-      </div>
+      <HelpSidebar />
     </div>
   )
 }
